Move tests from test_cortex to test_lib_snap
