+++ synapse/tests/test_lib_snap.py
@@ -1,3 +1,4 @@
+import os
 import synapse.exc as s_exc
 import synapse.lib.auth as s_auth
 import synapse.tests.common as s_t_common
@@ -89,3 +90,47 @@ class SnapTest(s_t_common.SynTest):
                         self.none(snap.getNodeByNdef(('testauto', 'bar')))
                         self.none(snap.getNodeByNdef(('testtime', 'baz')))
                         self.nn(snap.getNodeByNdef(('teststr', 'faz')))
+
+    def test_cortex_lift_layers_bad_filter(self):
+        '''
+        Test a two layer cortex where a lift operation gives the wrong result
+        '''
+        with self.getTestCore() as core1:
+            node = (('inet:ipv4', 1), {'props': {'asn': 42}, 'tags': {'woot': (None, None)}})
+            nodes_core1 = list(core1.addNodes([node]))
+
+            layerfn = os.path.join(core1.dirn, 'layers', 'default')
+            with self.getTestCore(conf={'layers': [layerfn]}) as core, core.snap() as snap:
+                # Basic sanity check
+                nodes = list(snap.getNodesBy('inet:ipv4', 1))
+                self.len(1, nodes)
+                self.eq(nodes_core1[0].pack(), nodes[0].pack())
+
+                # Now change asn in the "higher" layer
+                changed_node = (('inet:ipv4', 1), {'props': {'asn': 43}})
+                nodes = list(snap.addNodes([changed_node]))
+                nodes = list(snap.getNodesBy('inet:ipv4:asn', 42))
+                self.len(0, nodes)
+
+    def test_cortex_lift_layers_dup(self):
+        '''
+        Test a two layer cortex where a lift operation might give the same node twice incorrectly
+        '''
+        with self.getTestCore() as core1:
+            node = (('inet:ipv4', 1), {'props': {'asn': 42}})
+            nodes_core1 = list(core1.addNodes([node]))
+
+            layerfn = os.path.join(core1.dirn, 'layers', 'default')
+            with self.getTestCore(conf={'layers': [layerfn]}) as core, core.snap() as snap:
+                # Basic sanity check
+                nodes = list(snap.getNodesBy('inet:ipv4', 1))
+                self.len(1, nodes)
+                self.eq(nodes_core1[0].pack(), nodes[0].pack())
+
+                # Now set asn in the "higher" layer to the same (by changing it, then changing it back)
+                changed_node = (('inet:ipv4', 1), {'props': {'asn': 43}})
+                nodes = list(snap.addNodes([changed_node]))
+                changed_node = (('inet:ipv4', 1), {'props': {'asn': 42}})
+                nodes = list(snap.addNodes([changed_node]))
+                nodes = list(snap.getNodesBy('inet:ipv4:asn', 42))
+                self.len(1, nodes)
